@@ -44,8 +82,8 @@
     "css modules"
   ],
   "engines": {
-    "node": "4.2.4",
-    "npm": "3.8.8"
+    "node": "5.2.0",
+    "npm": "3.10.7"
   },
   "author": "Ryan Collins",
   "license": "MIT",
@@ -55,90 +93,115 @@
   "homepage": "https://github.com/RyanCCollins/scalable-react-boilerplate#readme",
   "dependencies": {
     "actions": "^1.3.0",
-    "babel-core": "^6.3.15",
-    "babel-loader": "^6.2.0",
-    "babel-polyfill": "latest",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
+    "apollo-client": "^0.5.20",
     "components": "^0.1.0",
+    "cross-env": "^2.0.0",
     "css-loader": "^0.23.0",
     "expect": "^1.20.2",
     "expect-jsx": "^2.6.0",
     "express": "^4.14.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
+    "enzyme": "^2.4.1",
+    "enzyme-to-json": "^1.0.3",
     "foundation-sites": "^6.2.3",
-    "grommet": "^0.6.10",
+    "functional-components": "0.0.5",
+    "graphql-tag": "^1.1.2",
+    "grommet-udacity": "v0.1.12",
     "history": "^1.14.0",
     "html-webpack-plugin": "^2.7.1",
     "immutability-helper": "^2.0.0",
     "immutable": "^3.7.5",
     "isomorphic-fetch": "^2.2.0",
+    "jest": "16.0.2",
+    "jest-cli": "16.0.2",
+    "jest-webpack-alias": "^3.3.3",
     "json-loader": "^0.5.4",
+    "markdown-loader": "^0.1.7",
     "minimist": "^1.2.0",
     "node-sass": "^3.4.2",
     "normalize.css": "^4.2.0",
     "postcss-loader": "^1.0.0",
+    "morgan": "^1.7.0",
+    "node-ensure": "0.0.0",
+    "offline-plugin": "^3.4.2",
     "react": "^15.1.0",
     "react-addons-css-transition-group": "^15.2.1",
+    "react-addons-test-utils": "^15.3.2",
+    "react-addons-update": "^15.3.2",
+    "react-apollo": "^0.5.13",
+    "react-css-modules": "^3.7.6",
     "react-dom": "^15.0.1",
     "react-foundation": "^0.6.8",
+    "react-hot-loader": "^3.0.0-beta.6",
     "react-intl": "^2.1.3",
     "react-redux": "^4.4.5",
     "react-router": "^2.3.0",
     "react-router-redux": "^4.0.4",
     "react-toolbox": "^1.2.1",
     "redux": "^3.5.2",
+    "redux-auth-wrapper": "^0.8.0",
     "redux-form": "^5.2.5",
     "redux-logger": "^2.6.1",
+    "redux-mock-store": "^1.1.4",
     "redux-observable": "^0.12.1",
     "redux-promise-middleware": "^3.2.0",
     "redux-thunk": "^1.0.0",
+    "reselect": "^2.5.4",
     "resolve-url-loader": "^1.4.4",
     "rxjs": "^5.0.0-rc.1",
     "sass-loader": "^3.1.2",
     "style-loader": "^0.13.0",
+    "serialize-javascript": "^1.3.0",
+    "styled-components": "latest",
+    "sinon": "2.0.0-pre",
     "svg-react-loader": "^0.3.6",
-    "webpack": "^1.12.9"
+    "utils": "^0.3.1"
   },
   "devDependencies": {
-    "babel-eslint": "^5.0.0-beta4",
+    "autoprefixer": "^6.5.1",
+    "babel-core": "^6.3.15",
+    "babel-eslint": "^6.0.4",
+    "babel-jest": "^16.0.0",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-webpack-alias": "^2.1.1",
+    "babel-plugin-webpack-loaders": "^0.7.1",
+    "babel-polyfill": "latest",
     "babel-preset-es2015": "^6.9.0",
-    "chai": "^3.4.1",
-    "chai-enzyme": "^0.5.0",
-    "chai-immutable": "^1.5.3",
-    "chai-jsx": "^1.0.1",
-    "cross-env": "^2.0.0",
-    "enzyme": "^2.4.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^4.0.0",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "babel-register": "^6.16.3",
+    "babel-runtime": "^6.11.6",
+    "css-loader": "^0.23.0",
+    "eslint": "^3.12.2",
+    "eslint-config-airbnb": "^13.0.0",
     "eslint-loader": "^1.1.1",
-    "eslint-plugin-react": "^3.11.3",
-    "expect-immutable": "0.0.3",
-    "isparta": "^4.0.0",
-    "isparta-loader": "2.0.0",
-    "jsdom": "^7.2.0",
-    "karma": "1.1.1",
-    "karma-chrome-launcher": "1.0.1",
-    "karma-coverage": "1.1.1",
-    "karma-firefox-launcher": "1.0.0",
-    "karma-ie-launcher": "1.0.0",
-    "karma-mocha": "1.1.1",
-    "karma-mocha-reporter": "2.0.4",
-    "karma-safari-launcher": "1.0.0",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "1.7.0",
-    "mocha": "^2.3.4",
+    "eslint-plugin-graphql": "^0.4.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-react": "^6.8.0",
+    "expect-jsx": "^2.6.0",
+    "extract-text-webpack-plugin": "^v2.0.0-beta.3",
+    "file-loader": "^0.9.0",
+    "html-loader": "^0.4.4",
+    "html-webpack-plugin": "^2.7.1",
+    "node-sass": "^3.4.2",
     "npm-install-webpack-plugin": "^4.0.3",
+    "expect-immutable": "0.0.3",
     "null-loader": "0.1.1",
     "plop": "1.5.0",
-    "react-css-modules": "^3.7.6",
-    "react-hot-loader": "^1.3.0",
+    "postcss-loader": "^1.1.0",
+    "precss": "^1.4.0",
+    "react-hot-loader": "^3.0.0-beta.6",
     "redux-devtools": "^3.0.1",
-    "redux-mock-store": "^1.1.4",
-    "sinon": "2.0.0-pre",
-    "webpack-dev-server": "^1.14.0",
+    "resolve-url-loader": "^1.4.4",
+    "rimraf": "^2.5.4",
+    "sass-loader": "^3.1.2",
+    "style-loader": "^0.13.0",
+    "webpack": "2.1.0-beta.19",
+    "webpack-dev-server": "2.1.0-beta.3",
+    "webpack-hot-middleware": "^2.10.0",
+    "webpack-manifest-plugin": "^1.1.0",
     "webpack-visualizer-plugin": "^0.1.5"
   }
 }
