@@ -1,3 +1,10 @@
-/* Assemble all components for export */
+/* GENERATOR: Assemble all components for export */
+export About from './About';
+export Contributor from './Contributor';
+export Avatar from './Avatar';
+export WelcomeModal from './WelcomeModal';
+export Divider from './Divider';
+export AppFooter from './AppFooter';
+export LoadingIndicator from './LoadingIndicator';
 export Navbar from './Navbar';
 export App from './App';
