@@ -1,2 +1,4 @@
-/* Assemble all containers for export */
-export FeatureFirstContainer from './FeatureFirstContainer';
+/* GENERATOR: Assemble all containers for export */
+export AboutContainer from './AboutContainer';
+export LandingContainer from './LandingContainer';
+export AppContainer from './AppContainer';
