@@ -1,8 +1,30 @@
-/* eslint-disable */ // React must be in scope here
 import React from 'react';
-/* eslint-enable */
 import { render } from 'react-dom';
-import routes from './routes';
+import { match } from 'react-router';
+import { install } from 'offline-plugin/runtime';
 import '../styles/styles.scss';
+import { history } from './store';
+import RouterApp, { routes } from './routes';
 
-render(routes, document.getElementById('app'));
+const isProduction = process.env.NODE_ENV === 'production';
+
+match({ history, routes },
+  (error, redirectLocation, renderProps) => { // eslint-disable-line
+    if (error) {
+      return console.error('Require.ensure error'); // eslint-disable-line
+    }
+    render(<RouterApp {...renderProps} />, document.getElementById('app'));
+  });
+
+if (isProduction) {
+  install();
+} else {
+  /* eslint-disable */
+  if (module.hot) {
+    module.hot.accept('./routes', () => {
+      const NewRouterApp = require('./routes').default;
+      render(<NewRouterApp />, document.getElementById('app'));
+    });
+  }
+  /* eslint-enable */
+}
