@@ -1,3 +1,4 @@
-/* Assemble all pages for export */
+/* GENERATOR: Assemble all pages for export */
+export AboutPage from './AboutPage/index';
 export NotFoundPage from './NotFoundPage/index';
 export LandingPage from './LandingPage/index';
