@@ -1,15 +1,19 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 import { reducer as formReducer } from 'redux-form';
+import client from './apolloClient';
 
-// Import all of your reducers here:
-import featureComponent from 'containers/FeatureFirstContainer/reducer';
+/* GENERATOR: Import all of your reducers */
+import landing from './containers/LandingContainer/reducer';
+import app from './containers/AppContainer/reducer';
 
 const rootReducer = combineReducers({
-  // Apply all of the reducers here.
-  featureComponent,
+  app,
+  /* GENERATOR: Compile all of your reducers */
+  landing,
   routing: routerReducer,
   form: formReducer,
+  apollo: client.reducer(),
 });
 
 export default rootReducer;
