@@ -1,24 +1,69 @@
+/* global System b:true */
 import React from 'react';
-import { Router, Route, IndexRoute } from 'react-router';
-import { Provider } from 'react-redux';
+import { Router } from 'react-router';
+import { ApolloProvider } from 'react-apollo';
 import store, { history } from './store';
+import client from './apolloClient';
+import { AppContainer } from 'containers'; // eslint-disable-line
+
 /* eslint-disable */
-import App from 'components/App';
-import * as Pages from 'pages';
+// Polyfill for the System.import
+if (typeof System === 'undefined') {
+  var System = {
+    import(path) {
+      return Promise.resolve(require(path));
+    },
+  };
+}
 /* eslint-enable */
 
-const routes = (
-  <Provider store={store}>
+// Switching to system.import to make use of dynamic tree shaking
+// https://medium.com/modus-create-front-end-development/automatic-code-splitting-for-react-router-w-es6-imports-a0abdaa491e9#.msrxv8fwd
+const errorLoading = err =>
+  console.error('Dynamic loading failed' + err); // eslint-disable-line
+
+const loadRoute = cb =>
+  module =>
+    cb(null, module.default);
+
+export const routes = {
+  component: AppContainer,
+  path: '/',
+  indexRoute: {
+    getComponent(location, callback) {
+      System.import('./pages/LandingPage') // eslint-disable-line block-scoped-var
+        .then(loadRoute(callback))
+        .catch(err => errorLoading(err));
+    },
+  },
+  childRoutes: [
+    {
+      path: '/about',
+      getComponent(location, callback) {
+        System.import('./pages/AboutPage') // eslint-disable-line block-scoped-var
+          .then(loadRoute(callback))
+          .catch(err => errorLoading(err));
+      },
+    },
+/* GENERATOR: Newly generated Routes go here */
+    {
+      path: '*',
+      getComponent(location, callback) {
+        System.import('./pages/NotFoundPage') // eslint-disable-line block-scoped-var
+          .then(loadRoute(callback))
+          .catch(err => errorLoading(err));
+      },
+    },
+  ],
+};
+
+const RouterApp = props => (
+  <ApolloProvider {...props} store={store} client={client}>
     <Router
-      history={history} // Scroll to top on route transitions
-      onUpdate={() => window.scrollTo(0, 0)} // eslint-disable-line
-    >
-      <Route path="/" component={App}>
-        <IndexRoute component={Pages.LandingPage} />
-        <Route path="*" component={Pages.NotFoundPage} />
-      </Route>
-    </Router>
-  </Provider>
+      history={history}
+      routes={routes}
+    />
+  </ApolloProvider>
 );
 
-export default routes;
+export default RouterApp;
