@@ -8,7 +8,7 @@ module.exports = {
       type: 'input',
       name: 'name',
       message: 'What should it be called?',
-      default: 'Scalable',
+      default: 'About',
       validate: value => {
         if ((/.+/).test(value)) {
           return componentNameCheck(value) ? 'A container with this name already exists' : true;
@@ -21,13 +21,31 @@ module.exports = {
       type: 'confirm',
       name: 'wantSCSSModules',
       default: true,
-      message: 'Does it need styling?',
+      message: 'Do you want to use SCSS Modules for styling?',
+    },
+    {
+      type: 'confirm',
+      name: 'wantStyledComponents',
+      default: false,
+      message: 'Do you want to use Styled Components for styling?',
     },
     {
       type: 'confirm',
       name: 'wantActionsAndReducer',
       default: true,
       message: 'Do you want actions/constants/reducer for this container?',
+    },
+    {
+      type: 'confirm',
+      name: 'wantSelectors',
+      default: false,
+      message: 'Do you want to use reselect?',
+    },
+    {
+      type: 'confirm',
+      name: 'wantGraphQL',
+      default: false,
+      message: 'Do you want a colocated GraphQL / Apollo query and mutation for this container?',
     },
   ],
   actions: (data) => {
@@ -47,7 +65,7 @@ module.exports = {
     actions.push({
       type: 'modify',
       path: '../../app/src/containers/index.js',
-      pattern: /(\/\* Assemble all containers for export \*\/)/g,
+      pattern: /(\/\* GENERATOR: Assemble all containers for export \*\/)/g,
       template: trimTemplateFile('config/generators/container/export.js.hbs'),
     });
 
@@ -56,6 +74,24 @@ module.exports = {
         type: 'add',
         path: '../../app/src/containers/{{properCase name}}Container/index.module.scss',
         templateFile: './container/styles.scss.hbs',
+        abortOnFail: true,
+      });
+    }
+    
+    if (data.wantStyledComponents) {
+      actions.push({
+        type: 'add',
+        path: '../../app/src/containers/{{properCase name}}Container/styles.js',
+        templateFile: './container/styles.js.hbs',
+        abortOnFail: true,
+      }) 
+    }
+
+    if (data.wantSelectors) {
+      actions.push({
+        type: 'add',
+        path: '../../app/src/containers/{{properCase name}}Container/selectors.js',
+        templateFile: './container/selectors.js.hbs',
         abortOnFail: true,
       });
     }
@@ -75,6 +111,34 @@ module.exports = {
         path: '../../app/src/containers/{{properCase name}}Container/tests/actions.test.js',
         templateFile: './container/actions.test.js.hbs',
         abortOnFail: true,
+      });
+
+      actions.push({
+        type: 'modify',
+        path: '../../app/src/store.js',
+        pattern: /(\/\* GENERATOR: Import all of your initial state \*\/)/g,
+        template: trimTemplateFile('config/generators/container/store.import.js.hbs'),
+      });
+
+      actions.push({
+        type: 'modify',
+        path: '../../app/src/store.js',
+        pattern: /(\/\* GENERATOR: Compile all of your initial state \*\/)/g,
+        template: trimTemplateFile('config/generators/container/store.usage.js.hbs'),
+      });
+
+      actions.push({
+        type: 'modify',
+        path: '../../app/src/reducers.js',
+        pattern: /(\/\* GENERATOR: Import all of your reducers \*\/)/g,
+        template: trimTemplateFile('config/generators/container/reducers.import.js.hbs'),
+      });
+
+      actions.push({
+        type: 'modify',
+        path: '../../app/src/reducers.js',
+        pattern: /(\/\* GENERATOR: Compile all of your reducers \*\/)/g,
+        template: trimTemplateFile('config/generators/container/reducers.usage.js.hbs'),
       });
 
       // README.md
